refactor(stable_api): extract RealtimeMixin into pyquotex._api.realtime
Moves 16 streaming/indicator methods from stable_api.py into a focused
mixin (start_*_stream, get_realtime_*, subscribe_indicator,
calculate_indicator, start_mood_stream, opening_closing_current_candle,
get_signal_data). Public surface unchanged.
